refactor: :recycle: Separated the templatae service into: renderTemplateService, revertTemplateService and templateServiceUtils
From now the create commands will depend on renderTemplateService, delete commands will depend on revertTemplateService and both of these new services will depend on templateServiceUtils.
