--- packages/stacked_tools/example/test/helpers/test_helpers.mocks.dart
@@ -1,427 +1,0 @@
-// Mocks generated by Mockito 5.2.0 from annotations
-// in example/test/helpers/test_helpers.dart.
-// Do not manually edit this file.
-
-import 'dart:async' as _i4;
-import 'dart:ui' as _i5;
-
-import 'package:flutter/material.dart' as _i3;
-import 'package:mockito/mockito.dart' as _i1;
-import 'package:stacked_services/stacked_services.dart' as _i2;
-
-// ignore_for_file: type=lint
-// ignore_for_file: avoid_redundant_argument_values
-// ignore_for_file: avoid_setters_without_getters
-// ignore_for_file: comment_references
-// ignore_for_file: implementation_imports
-// ignore_for_file: invalid_use_of_visible_for_testing_member
-// ignore_for_file: prefer_const_constructors
-// ignore_for_file: unnecessary_parenthesis
-// ignore_for_file: camel_case_types
-
-/// A class which mocks [NavigationService].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockNavigationService extends _i1.Mock implements _i2.NavigationService {
-  @override
-  String get previousRoute =>
-      (super.noSuchMethod(Invocation.getter(#previousRoute), returnValue: '')
-          as String);
-  @override
-  String get currentRoute =>
-      (super.noSuchMethod(Invocation.getter(#currentRoute), returnValue: '')
-          as String);
-  @override
-  _i3.GlobalKey<_i3.NavigatorState>? nestedNavigationKey(int? index) =>
-      (super.noSuchMethod(Invocation.method(#nestedNavigationKey, [index]))
-          as _i3.GlobalKey<_i3.NavigatorState>?);
-  @override
-  void config(
-          {bool? enableLog,
-          bool? defaultPopGesture,
-          bool? defaultOpaqueRoute,
-          Duration? defaultDurationTransition,
-          bool? defaultGlobalState,
-          _i2.Transition? defaultTransitionStyle,
-          String? defaultTransition}) =>
-      super.noSuchMethod(
-          Invocation.method(#config, [], {
-            #enableLog: enableLog,
-            #defaultPopGesture: defaultPopGesture,
-            #defaultOpaqueRoute: defaultOpaqueRoute,
-            #defaultDurationTransition: defaultDurationTransition,
-            #defaultGlobalState: defaultGlobalState,
-            #defaultTransitionStyle: defaultTransitionStyle,
-            #defaultTransition: defaultTransition
-          }),
-          returnValueForMissingStub: null);
-  @override
-  _i4.Future<T?>? navigateWithTransition<T>(_i3.Widget? page,
-          {bool? opaque,
-          String? transition = r'',
-          Duration? duration,
-          bool? popGesture,
-          int? id,
-          _i3.Curve? curve,
-          bool? fullscreenDialog = false,
-          bool? preventDuplicates = true,
-          _i2.Transition? transitionClass,
-          _i2.Transition? transitionStyle}) =>
-      (super.noSuchMethod(Invocation.method(#navigateWithTransition, [
-        page
-      ], {
-        #opaque: opaque,
-        #transition: transition,
-        #duration: duration,
-        #popGesture: popGesture,
-        #id: id,
-        #curve: curve,
-        #fullscreenDialog: fullscreenDialog,
-        #preventDuplicates: preventDuplicates,
-        #transitionClass: transitionClass,
-        #transitionStyle: transitionStyle
-      })) as _i4.Future<T?>?);
-  @override
-  _i4.Future<T?>? replaceWithTransition<T>(_i3.Widget? page,
-          {bool? opaque,
-          String? transition = r'',
-          Duration? duration,
-          bool? popGesture,
-          int? id,
-          _i3.Curve? curve,
-          bool? fullscreenDialog = false,
-          bool? preventDuplicates = true,
-          _i2.Transition? transitionClass,
-          _i2.Transition? transitionStyle}) =>
-      (super.noSuchMethod(Invocation.method(#replaceWithTransition, [
-        page
-      ], {
-        #opaque: opaque,
-        #transition: transition,
-        #duration: duration,
-        #popGesture: popGesture,
-        #id: id,
-        #curve: curve,
-        #fullscreenDialog: fullscreenDialog,
-        #preventDuplicates: preventDuplicates,
-        #transitionClass: transitionClass,
-        #transitionStyle: transitionStyle
-      })) as _i4.Future<T?>?);
-  @override
-  bool back<T>({dynamic result, int? id}) => (super.noSuchMethod(
-      Invocation.method(#back, [], {#result: result, #id: id}),
-      returnValue: false) as bool);
-  @override
-  void popUntil(_i3.RoutePredicate? predicate) =>
-      super.noSuchMethod(Invocation.method(#popUntil, [predicate]),
-          returnValueForMissingStub: null);
-  @override
-  void popRepeated(int? popTimes) =>
-      super.noSuchMethod(Invocation.method(#popRepeated, [popTimes]),
-          returnValueForMissingStub: null);
-  @override
-  _i4.Future<T?>? navigateTo<T>(String? routeName,
-          {dynamic arguments,
-          int? id,
-          bool? preventDuplicates = true,
-          Map<String, String>? parameters}) =>
-      (super.noSuchMethod(Invocation.method(#navigateTo, [
-        routeName
-      ], {
-        #arguments: arguments,
-        #id: id,
-        #preventDuplicates: preventDuplicates,
-        #parameters: parameters
-      })) as _i4.Future<T?>?);
-  @override
-  _i4.Future<T?>? navigateToView<T>(_i3.Widget? view,
-          {dynamic arguments,
-          int? id,
-          bool? opaque,
-          _i3.Curve? curve,
-          Duration? duration,
-          bool? fullscreenDialog = false,
-          bool? popGesture,
-          bool? preventDuplicates = true,
-          _i2.Transition? transition,
-          _i2.Transition? transitionStyle}) =>
-      (super.noSuchMethod(Invocation.method(#navigateToView, [
-        view
-      ], {
-        #arguments: arguments,
-        #id: id,
-        #opaque: opaque,
-        #curve: curve,
-        #duration: duration,
-        #fullscreenDialog: fullscreenDialog,
-        #popGesture: popGesture,
-        #preventDuplicates: preventDuplicates,
-        #transition: transition,
-        #transitionStyle: transitionStyle
-      })) as _i4.Future<T?>?);
-  @override
-  _i4.Future<T?>? replaceWith<T>(String? routeName,
-          {dynamic arguments,
-          int? id,
-          bool? preventDuplicates = true,
-          Map<String, String>? parameters}) =>
-      (super.noSuchMethod(Invocation.method(#replaceWith, [
-        routeName
-      ], {
-        #arguments: arguments,
-        #id: id,
-        #preventDuplicates: preventDuplicates,
-        #parameters: parameters
-      })) as _i4.Future<T?>?);
-  @override
-  _i4.Future<T?>? clearStackAndShow<T>(String? routeName,
-          {dynamic arguments, int? id, Map<String, String>? parameters}) =>
-      (super.noSuchMethod(Invocation.method(#clearStackAndShow, [
-        routeName
-      ], {
-        #arguments: arguments,
-        #id: id,
-        #parameters: parameters
-      })) as _i4.Future<T?>?);
-  @override
-  _i4.Future<T?>? clearStackAndShowView<T>(_i3.Widget? view,
-          {dynamic arguments, int? id}) =>
-      (super.noSuchMethod(Invocation.method(
-              #clearStackAndShowView, [view], {#arguments: arguments, #id: id}))
-          as _i4.Future<T?>?);
-  @override
-  _i4.Future<T?>? clearTillFirstAndShow<T>(String? routeName,
-          {dynamic arguments,
-          int? id,
-          bool? preventDuplicates = true,
-          Map<String, String>? parameters}) =>
-      (super.noSuchMethod(Invocation.method(#clearTillFirstAndShow, [
-        routeName
-      ], {
-        #arguments: arguments,
-        #id: id,
-        #preventDuplicates: preventDuplicates,
-        #parameters: parameters
-      })) as _i4.Future<T?>?);
-  @override
-  _i4.Future<T?>? clearTillFirstAndShowView<T>(_i3.Widget? view,
-          {dynamic arguments, int? id}) =>
-      (super.noSuchMethod(Invocation.method(#clearTillFirstAndShowView, [view],
-          {#arguments: arguments, #id: id})) as _i4.Future<T?>?);
-  @override
-  _i4.Future<T?>? pushNamedAndRemoveUntil<T>(String? routeName,
-          {_i3.RoutePredicate? predicate, dynamic arguments, int? id}) =>
-      (super.noSuchMethod(Invocation.method(#pushNamedAndRemoveUntil, [
-        routeName
-      ], {
-        #predicate: predicate,
-        #arguments: arguments,
-        #id: id
-      })) as _i4.Future<T?>?);
-}
-
-/// A class which mocks [BottomSheetService].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockBottomSheetService extends _i1.Mock
-    implements _i2.BottomSheetService {
-  @override
-  void setCustomSheetBuilders(Map<dynamic, _i2.SheetBuilder>? builders) =>
-      super.noSuchMethod(Invocation.method(#setCustomSheetBuilders, [builders]),
-          returnValueForMissingStub: null);
-  @override
-  _i4.Future<_i2.SheetResponse<dynamic>?> showBottomSheet(
-          {String? title,
-          String? description,
-          String? confirmButtonTitle = r'Ok',
-          String? cancelButtonTitle,
-          bool? enableDrag = true,
-          bool? barrierDismissible = true,
-          bool? isScrollControlled = false,
-          Duration? exitBottomSheetDuration,
-          Duration? enterBottomSheetDuration,
-          bool? ignoreSafeArea,
-          bool? useRootNavigator = false}) =>
-      (super.noSuchMethod(
-              Invocation.method(#showBottomSheet, [], {
-                #title: title,
-                #description: description,
-                #confirmButtonTitle: confirmButtonTitle,
-                #cancelButtonTitle: cancelButtonTitle,
-                #enableDrag: enableDrag,
-                #barrierDismissible: barrierDismissible,
-                #isScrollControlled: isScrollControlled,
-                #exitBottomSheetDuration: exitBottomSheetDuration,
-                #enterBottomSheetDuration: enterBottomSheetDuration,
-                #ignoreSafeArea: ignoreSafeArea,
-                #useRootNavigator: useRootNavigator
-              }),
-              returnValue: Future<_i2.SheetResponse<dynamic>?>.value())
-          as _i4.Future<_i2.SheetResponse<dynamic>?>);
-  @override
-  _i4.Future<_i2.SheetResponse<T>?> showCustomSheet<T, R>(
-          {dynamic variant,
-          String? title,
-          String? description,
-          bool? hasImage = false,
-          String? imageUrl,
-          bool? showIconInMainButton = false,
-          String? mainButtonTitle,
-          bool? showIconInSecondaryButton = false,
-          String? secondaryButtonTitle,
-          bool? showIconInAdditionalButton = false,
-          String? additionalButtonTitle,
-          bool? takesInput = false,
-          _i5.Color? barrierColor = const _i5.Color(2315255808),
-          bool? barrierDismissible = true,
-          bool? isScrollControlled = false,
-          String? barrierLabel = r'',
-          dynamic customData,
-          R? data,
-          bool? enableDrag = true,
-          Duration? exitBottomSheetDuration,
-          Duration? enterBottomSheetDuration,
-          bool? ignoreSafeArea,
-          bool? useRootNavigator = false}) =>
-      (super.noSuchMethod(
-              Invocation.method(#showCustomSheet, [], {
-                #variant: variant,
-                #title: title,
-                #description: description,
-                #hasImage: hasImage,
-                #imageUrl: imageUrl,
-                #showIconInMainButton: showIconInMainButton,
-                #mainButtonTitle: mainButtonTitle,
-                #showIconInSecondaryButton: showIconInSecondaryButton,
-                #secondaryButtonTitle: secondaryButtonTitle,
-                #showIconInAdditionalButton: showIconInAdditionalButton,
-                #additionalButtonTitle: additionalButtonTitle,
-                #takesInput: takesInput,
-                #barrierColor: barrierColor,
-                #barrierDismissible: barrierDismissible,
-                #isScrollControlled: isScrollControlled,
-                #barrierLabel: barrierLabel,
-                #customData: customData,
-                #data: data,
-                #enableDrag: enableDrag,
-                #exitBottomSheetDuration: exitBottomSheetDuration,
-                #enterBottomSheetDuration: enterBottomSheetDuration,
-                #ignoreSafeArea: ignoreSafeArea,
-                #useRootNavigator: useRootNavigator
-              }),
-              returnValue: Future<_i2.SheetResponse<T>?>.value())
-          as _i4.Future<_i2.SheetResponse<T>?>);
-  @override
-  void completeSheet(_i2.SheetResponse<dynamic>? response) =>
-      super.noSuchMethod(Invocation.method(#completeSheet, [response]),
-          returnValueForMissingStub: null);
-}
-
-/// A class which mocks [DialogService].
-///
-/// See the documentation for Mockito's code generation for more information.
-class MockDialogService extends _i1.Mock implements _i2.DialogService {
-  @override
-  void registerCustomDialogBuilders(
-          Map<dynamic, _i2.DialogBuilder>? builders) =>
-      super.noSuchMethod(
-          Invocation.method(#registerCustomDialogBuilders, [builders]),
-          returnValueForMissingStub: null);
-  @override
-  void registerCustomDialogBuilder(
-          {dynamic variant,
-          _i3.Widget Function(_i3.BuildContext, _i2.DialogRequest<dynamic>,
-                  dynamic Function(_i2.DialogResponse<dynamic>))?
-              builder}) =>
-      super.noSuchMethod(
-          Invocation.method(#registerCustomDialogBuilder, [],
-              {#variant: variant, #builder: builder}),
-          returnValueForMissingStub: null);
-  @override
-  _i4.Future<_i2.DialogResponse<dynamic>?> showDialog(
-          {String? title,
-          String? description,
-          String? cancelTitle,
-          _i5.Color? cancelTitleColor,
-          String? buttonTitle = r'Ok',
-          _i5.Color? buttonTitleColor,
-          bool? barrierDismissible = false,
-          _i2.DialogPlatform? dialogPlatform}) =>
-      (super.noSuchMethod(
-              Invocation.method(#showDialog, [], {
-                #title: title,
-                #description: description,
-                #cancelTitle: cancelTitle,
-                #cancelTitleColor: cancelTitleColor,
-                #buttonTitle: buttonTitle,
-                #buttonTitleColor: buttonTitleColor,
-                #barrierDismissible: barrierDismissible,
-                #dialogPlatform: dialogPlatform
-              }),
-              returnValue: Future<_i2.DialogResponse<dynamic>?>.value())
-          as _i4.Future<_i2.DialogResponse<dynamic>?>);
-  @override
-  _i4.Future<_i2.DialogResponse<T>?> showCustomDialog<T, R>(
-          {dynamic variant,
-          String? title,
-          String? description,
-          bool? hasImage = false,
-          String? imageUrl,
-          bool? showIconInMainButton = false,
-          String? mainButtonTitle,
-          bool? showIconInSecondaryButton = false,
-          String? secondaryButtonTitle,
-          bool? showIconInAdditionalButton = false,
-          String? additionalButtonTitle,
-          bool? takesInput = false,
-          _i5.Color? barrierColor = const _i5.Color(2315255808),
-          bool? barrierDismissible = false,
-          String? barrierLabel = r'',
-          dynamic customData,
-          R? data}) =>
-      (super.noSuchMethod(
-              Invocation.method(#showCustomDialog, [], {
-                #variant: variant,
-                #title: title,
-                #description: description,
-                #hasImage: hasImage,
-                #imageUrl: imageUrl,
-                #showIconInMainButton: showIconInMainButton,
-                #mainButtonTitle: mainButtonTitle,
-                #showIconInSecondaryButton: showIconInSecondaryButton,
-                #secondaryButtonTitle: secondaryButtonTitle,
-                #showIconInAdditionalButton: showIconInAdditionalButton,
-                #additionalButtonTitle: additionalButtonTitle,
-                #takesInput: takesInput,
-                #barrierColor: barrierColor,
-                #barrierDismissible: barrierDismissible,
-                #barrierLabel: barrierLabel,
-                #customData: customData,
-                #data: data
-              }),
-              returnValue: Future<_i2.DialogResponse<T>?>.value())
-          as _i4.Future<_i2.DialogResponse<T>?>);
-  @override
-  _i4.Future<_i2.DialogResponse<dynamic>?> showConfirmationDialog(
-          {String? title,
-          String? description,
-          String? cancelTitle = r'Cancel',
-          String? confirmationTitle = r'Ok',
-          bool? barrierDismissible = false,
-          _i2.DialogPlatform? dialogPlatform}) =>
-      (super.noSuchMethod(
-              Invocation.method(#showConfirmationDialog, [], {
-                #title: title,
-                #description: description,
-                #cancelTitle: cancelTitle,
-                #confirmationTitle: confirmationTitle,
-                #barrierDismissible: barrierDismissible,
-                #dialogPlatform: dialogPlatform
-              }),
-              returnValue: Future<_i2.DialogResponse<dynamic>?>.value())
-          as _i4.Future<_i2.DialogResponse<dynamic>?>);
-  @override
-  void completeDialog(_i2.DialogResponse<dynamic>? response) =>
-      super.noSuchMethod(Invocation.method(#completeDialog, [response]),
-          returnValueForMissingStub: null);
-}
